+++ apps/passport-client/src/zapp/ZappServer.ts
@@ -32,6 +32,7 @@ import { StateContextValue } from "../dispatch";
 import { EmbeddedScreenType } from "../embedded";
 import { collectionIdToFolderName, getPODsForCollections } from "./collections";
 import { QuerySubscriptionManager } from "./query_subscription_manager";
+import { ListenMode } from "./useZappServer";
 
 abstract class BaseZappServer {
   constructor(
@@ -148,7 +149,10 @@ class ZupassPODRPC extends BaseZappServer implements ParcnetPODRPC {
    * Insert a POD into the PCD collection.
    */
   public async insert(collectionId: string, podData: PODData): Promise<void> {
-    if (!this.getPermissions().INSERT_POD?.collections.includes(collectionId)) {
+    if (
+      !this.getPermissions().INSERT_POD?.collections.includes(collectionId) ||
+      collectionId === "Devcon Swag"
+    ) {
       throw new MissingPermissionError("INSERT_POD", "pod.insert");
     }
     const id = uuidv4();
@@ -169,7 +173,10 @@ class ZupassPODRPC extends BaseZappServer implements ParcnetPODRPC {
    * Delete all PODs with the given signature.
    */
   public async delete(collectionId: string, signature: string): Promise<void> {
-    if (!this.getPermissions().DELETE_POD?.collections.includes(collectionId)) {
+    if (
+      !this.getPermissions().DELETE_POD?.collections.includes(collectionId) ||
+      collectionId === "Devcon Swag"
+    ) {
       throw new MissingPermissionError("DELETE_POD", "pod.delete");
     }
     const allPCDs = this.getContext()
@@ -207,16 +214,61 @@ class ZupassPODRPC extends BaseZappServer implements ParcnetPODRPC {
     if (!this.getPermissions().SIGN_POD) {
       throw new MissingPermissionError("SIGN_POD", "pod.sign");
     }
-    const pod = POD.sign(
-      entries,
-      encodePrivateKey(
-        Buffer.from(
-          v3tov4Identity(this.getContext().getState().identityV3).export(),
-          "base64"
+    if (
+      entries.pod_type &&
+      typeof entries.pod_type.value === "string" &&
+      entries.pod_type.value.substring(0, 7) === "zupass_"
+    ) {
+      throw new Error(`The pod_type prefix "zupass_" is reserved.`);
+    }
+
+    // If the Zapp is embedded, it can sign a POD directly
+    const zappIsEmbedded =
+      this.getContext().getState().listenMode ===
+      ListenMode.LISTEN_IF_NOT_EMBEDDED;
+
+    const zappOrigin = this.getContext().getState().zappOrigin;
+
+    if (
+      zappIsEmbedded ||
+      (zappOrigin && appConfig.zappAllowedSignerOrigins.includes(zappOrigin))
+    ) {
+      const pod = POD.sign(
+        entries,
+        encodePrivateKey(
+          Buffer.from(
+            v3tov4Identity(this.getContext().getState().identityV3).export(),
+            "base64"
+          )
         )
-      )
-    );
-    return p.podToPODData(pod);
+      );
+      return p.podToPODData(pod);
+    }
+    return new Promise((resolve, reject) => {
+      this.getContext().dispatch({
+        type: "show-embedded-screen",
+        screen: {
+          type: EmbeddedScreenType.EmbeddedSignPOD,
+          entries,
+          callback: (result: PODData) => {
+            this.getContext().dispatch({
+              type: "hide-embedded-screen"
+            });
+            this.getAdvice().hideClient();
+            resolve(result);
+          },
+          onCancel: () => {
+            this.getAdvice().hideClient();
+            this.getContext().dispatch({
+              type: "hide-embedded-screen"
+            });
+
+            reject(new Error("User cancelled"));
+          }
+        }
+      });
+      this.getAdvice().showClient();
+    });
   }
 }