+++ apps/passport-server/src/database/queries/saveUser.ts
@@ -1,6 +1,6 @@
-import { Pool } from "postgres-pool";
+import { PoolClient } from "postgres-pool";
 import { logger } from "../../util/logger";
-import { sqlQuery, sqlTransaction } from "../sqlQuery";
+import { sqlQuery } from "../sqlQuery";
 
 export interface SaveUserParams {
   uuid: string;
@@ -22,7 +22,7 @@ export interface SaveUserParams {
  * @param params is a partial {@link UserRow}.
  */
 export async function upsertUser(
-  client: Pool,
+  client: PoolClient,
   params: SaveUserParams
 ): Promise<string> {
   logger("[DB] upsertUser", params);
@@ -37,25 +37,21 @@ export async function upsertUser(
     );
   }
 
-  return sqlTransaction(
-    client,
-    "save user",
-    async (client) => {
-      const {
-        commitment,
-        salt,
-        encryption_key,
-        terms_agreed,
-        extra_issuance,
-        uuid,
-        emails,
-        semaphore_v4_commitment,
-        semaphore_v4_pubkey
-      } = params;
+  const {
+    commitment,
+    salt,
+    encryption_key,
+    terms_agreed,
+    extra_issuance,
+    uuid,
+    emails,
+    semaphore_v4_commitment,
+    semaphore_v4_pubkey
+  } = params;
 
-      const upsertUserResult = await sqlQuery(
-        client,
-        `\
+  const upsertUserResult = await sqlQuery(
+    client,
+    `\
 INSERT INTO users 
 (uuid, commitment, salt, encryption_key, terms_agreed, extra_issuance, semaphore_v4_commitment, semaphore_v4_pubkey)
 VALUES 
@@ -63,46 +59,43 @@ VALUES
 ON CONFLICT (uuid) DO UPDATE SET 
 commitment = $2, salt = $3, encryption_key = $4, terms_agreed = $5, extra_issuance=$6, time_updated=$7, semaphore_v4_commitment=$8, semaphore_v4_pubkey=$9
 returning *`,
-        [
-          uuid,
-          commitment,
-          salt,
-          encryption_key,
-          terms_agreed,
-          extra_issuance,
-          new Date(),
-          semaphore_v4_commitment,
-          semaphore_v4_pubkey
-        ],
-        0
-      );
-
-      const user = upsertUserResult.rows[0];
-      if (!user) {
-        throw new Error(`Failed to save user.`);
-      }
-
-      // Update the user's associated emails
-      await sqlQuery(
-        client,
-        `DELETE FROM user_emails WHERE user_id = $1`,
-        [uuid],
-        0
-      );
-
-      const emailValues = emails
-        .map((_, index) => `($1, $${index + 2})`)
-        .join(", ");
-
-      await sqlQuery(
-        client,
-        `INSERT INTO user_emails (user_id, email) VALUES ${emailValues}`,
-        [uuid, ...emails],
-        0
-      );
-
-      return uuid;
-    },
-    3
+    [
+      uuid,
+      commitment,
+      salt,
+      encryption_key,
+      terms_agreed,
+      extra_issuance,
+      new Date(),
+      semaphore_v4_commitment,
+      semaphore_v4_pubkey
+    ],
+    0
   );
+
+  const user = upsertUserResult.rows[0];
+  if (!user) {
+    throw new Error(`Failed to save user.`);
+  }
+
+  // Update the user's associated emails
+  await sqlQuery(
+    client,
+    `DELETE FROM user_emails WHERE user_id = $1`,
+    [uuid],
+    0
+  );
+
+  const emailValues = emails
+    .map((_, index) => `($1, $${index + 2})`)
+    .join(", ");
+
+  await sqlQuery(
+    client,
+    `INSERT INTO user_emails (user_id, email) VALUES ${emailValues}`,
+    [uuid, ...emails],
+    0
+  );
+
+  return uuid;
 }