+++ packages/pcd/pod-pcd/src/PODPCDPackage.ts
@@ -1,5 +1,5 @@
 import { DisplayOptions, PCD, PCDPackage, SerializedPCD } from "@pcd/pcd-types";
-import { POD } from "@pcd/pod";
+import { POD, podEntriesFromJSON, requireType } from "@pcd/pod";
 import { requireDefinedParameter } from "@pcd/util";
 import JSONBig from "json-bigint";
 import { v4 as uuid } from "uuid";
@@ -11,6 +11,7 @@ import {
   PODPCDProof,
   PODPCDTypeName
 } from "./PODPCD";
+import { getTitleEntry } from "./utils";
 
 /**
  * Creates a new {@link PODPCD} by generating an {@link PODPCDProof}
@@ -23,7 +24,10 @@ export async function prove(args: PODPCDArgs): Promise<PODPCD> {
   if (!args.entries.value) throw new Error("No POD entries value provided");
   const id = typeof args.id.value === "string" ? args.id.value : uuid();
 
-  const pod = POD.sign(args.entries.value, args.privateKey.value);
+  const pod = POD.sign(
+    podEntriesFromJSON(args.entries.value),
+    args.privateKey.value
+  );
   return new PODPCD(id, pod);
 }
 
@@ -44,13 +48,9 @@ export async function verify(pcd: PODPCD): Promise<boolean> {
 export async function serialize(pcd: PODPCD): Promise<SerializedPCD<PODPCD>> {
   return {
     type: PODPCDTypeName,
-    pcd: JSONBig({
-      useNativeBigInt: true,
-      alwaysParseAsBig: true
-    }).stringify({
+    pcd: JSON.stringify({
       id: pcd.id,
-      claim: pcd.claim,
-      proof: pcd.proof
+      jsonPOD: pcd.pod.toJSON()
     })
   };
 }
@@ -61,14 +61,38 @@ export async function serialize(pcd: PODPCD): Promise<SerializedPCD<PODPCD>> {
  * @returns The deserialized version of the POD PCD.
  */
 export async function deserialize(serialized: string): Promise<PODPCD> {
+  try {
+    const deserialized = JSON.parse(serialized);
+    requireType("id", deserialized.id, "string");
+    requireType("jsonPOD", deserialized.jsonPOD, "object");
+    const pod = POD.fromJSON(deserialized.jsonPOD);
+    return new PODPCD(deserialized.id, pod);
+  } catch (jsonError) {
+    // Try the old legacy JSONBigint format for backward compatibility.
+    // Its JSON(ish) structure contained `id`, `claim`, and `proof` so
+    // isn't mistakable for the new format with `jsonPOD` used above.
+    try {
+      return deserializeLegacyJSONBigint(serialized);
+    } catch (legacyError) {
+      // Old format didn't work either, so throw the original exception.
+      throw jsonError;
+    }
+  }
+}
+
+async function deserializeLegacyJSONBigint(
+  serialized: string
+): Promise<PODPCD> {
+  // Legacy format which predates JSON encoding used json-bigint to directly
+  // serialize PODEntries.  This was eliminated due to lack of composability,
+  // but there are Zupass users who already had PODs which they may need to
+  // load.  They'll get upgraded when they re-serialize, but we can't ever
+  // guarantee they're all gone since long-idle users might return.
   const deserialized = JSONBig({
     useNativeBigInt: true,
     alwaysParseAsBig: true
   }).parse(serialized);
 
-  // TODO(POD-P2): More careful schema validation, likely with Zod, with
-  // special handling of the PODEntries type and subtypes.
-  // TODO(POD-P3): Backward-compatible schema versioning.
   requireDefinedParameter(deserialized.id, "id");
   requireDefinedParameter(deserialized.claim, "claim");
   requireDefinedParameter(deserialized.claim.entries, "entries");
@@ -100,7 +124,7 @@ export function getDisplayOptions(
   // data outside of claim + proof?
   pcd: PCD<PODPCDClaim, PODPCDProof>
 ): DisplayOptions {
-  const titleEntry = pcd.claim.entries["zupass_title"];
+  const titleEntry = getTitleEntry(pcd);
   if (titleEntry?.type === "string" && titleEntry.value.length > 0) {
     return {
       header: titleEntry.value,