+++ packages/patched/snarkjs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @pcd/snarkjs
 
+## 0.9.0
+
+### Minor Changes
+
+- Release for compatibility with Zupass prod
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/circom_runtime@0.3.0
+
+## 0.8.0
+
+### Minor Changes
+
+- Compatibility with updated Zupass with new GPC artifacts
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/circom_runtime@0.2.0
+
 ## 0.7.7
 
 ### Patch Changes