+++ packages/pcd/halo-nonce-pcd/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @pcd/halo-nonce-pcd
 
+## 0.9.0
+
+### Minor Changes
+
+- Release for compatibility with Zupass prod
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/pcd-types@0.13.0
+  - @pcd/util@0.7.0
+
+## 0.8.0
+
+### Minor Changes
+
+- Compatibility with updated Zupass with new GPC artifacts
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/pcd-types@0.12.0
+  - @pcd/util@0.6.0
+
 ## 0.7.5
 
 ### Patch Changes