+++ packages/pcd/obj-pcd/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @pcd/obj-pcd
 
+## 0.2.0
+
+### Minor Changes
+
+- Release for compatibility with Zupass prod
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/pcd-types@0.13.0
+  - @pcd/util@0.7.0
+
+## 0.1.0
+
+### Minor Changes
+
+- Compatibility with updated Zupass with new GPC artifacts
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/pcd-types@0.12.0
+  - @pcd/util@0.6.0
+
 ## 0.0.5
 
 ### Patch Changes