+++ packages/ui/eddsa-frog-pcd-ui/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @pcd/eddsa-frog-pcd-ui
 
+## 0.4.0
+
+### Minor Changes
+
+- Release for compatibility with Zupass prod
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/eddsa-frog-pcd@0.5.0
+  - @pcd/passport-ui@0.13.0
+  - @pcd/pcd-types@0.13.0
+
+## 0.3.0
+
+### Minor Changes
+
+- Compatibility with updated Zupass with new GPC artifacts
+
+### Patch Changes
+
+- Updated dependencies
+  - @pcd/passport-ui@0.12.0
+  - @pcd/pcd-types@0.12.0
+  - @pcd/eddsa-frog-pcd@0.4.0
+
 ## 0.2.6
 
 ### Patch Changes